DotPy_task/requirements.txt
numpy==2.4.2
opencv_python==4.13.0.92
opencv_python_headless==4.13.0.92
Pillow==12.1.0
streamlit==1.54.0
ultralytics==8.4.12
